Fix ALL tests
