+++ src/middleware/tests/.env
@@ -3,6 +3,7 @@ SEMAPPS_HOME_URL=http://localhost:3000/
 # Fuseki instance with ACL
 SEMAPPS_SPARQL_ENDPOINT=http://localhost:3040/
 SEMAPPS_MAIN_DATASET=testData
+SEMAPPS_SETTINGS_DATASET=settings_test
 SEMAPPS_JENA_USER=admin
 SEMAPPS_JENA_PASSWORD=admin
