+++ src/middleware/tests/config.js
@@ -6,6 +6,7 @@ module.exports = {
   HOME_URL: process.env.SEMAPPS_HOME_URL,
   SPARQL_ENDPOINT: process.env.SEMAPPS_SPARQL_ENDPOINT,
   MAIN_DATASET: process.env.SEMAPPS_MAIN_DATASET,
+  SETTINGS_DATASET: process.env.SEMAPPS_SETTINGS_DATASET,
   JENA_USER: process.env.SEMAPPS_JENA_USER,
   JENA_PASSWORD: process.env.SEMAPPS_JENA_PASSWORD
 };
