+++ src/middleware/tests/webId/webId.test.js
@@ -1,14 +1,12 @@
 const { ServiceBroker } = require('moleculer');
 const { CoreService } = require('@semapps/core');
 const { WebIdService } = require('@semapps/webid');
-const { WebAclMiddleware } = require('@semapps/webacl');
-const EventsWatcher = require('../middleware/EventsWatcher');
 const CONFIG = require('../config');
 const ontologies = require('../ontologies');
+const path = require("path");
 
 jest.setTimeout(20000);
 const broker = new ServiceBroker({
-  middlewares: [EventsWatcher, WebAclMiddleware],
   logger: {
     type: 'Console',
     options: {
@@ -21,6 +19,7 @@ beforeAll(async () => {
   await broker.createService(CoreService, {
     settings: {
       baseUrl: CONFIG.HOME_URL,
+      baseDir: path.resolve(__dirname, '..'),
       triplestore: {
         url: CONFIG.SPARQL_ENDPOINT,
         user: CONFIG.JENA_USER,
@@ -28,7 +27,12 @@ beforeAll(async () => {
         mainDataset: CONFIG.MAIN_DATASET
       },
       ontologies,
-      containers: ['/users']
+      containers: ['/users'],
+      activitypub: false,
+      mirror: false,
+      void: false,
+      webacl: false,
+      webfinger: false,
     }
   });
   broker.createService(WebIdService, {
@@ -42,17 +46,6 @@ beforeAll(async () => {
   await broker.call('triplestore.dropAll', { webId: 'system' });
   await broker.stop();
   await broker.start();
-
-  // setting some write permission on the container for anonymous user, which is the one that will be used in the tests.
-  await broker.call('webacl.resource.addRights', {
-    resourceUri: CONFIG.HOME_URL + 'users',
-    additionalRights: {
-      anon: {
-        write: true
-      }
-    },
-    webId: 'system'
-  });
 });
 
 afterAll(async () => {
